Initial refactor into single codepath for PR
