+++ Assets/Scripts/Terrain/TerrainTexturing.cs
@@ -4,18 +4,16 @@
 // License:         MIT License (http://www.opensource.org/licenses/mit-license.php)
 // Source Code:     https://github.com/Interkarma/daggerfall-unity
 // Original Author: Gavin Clayton (interkarma@dfworkshop.net)
-// Contributors:    
+// Contributors:    Hazelnut
 // 
 // Notes:
 //
 
 using UnityEngine;
 using System;
-using System.Collections;
-using System.Collections.Generic;
-using DaggerfallConnect;
 using DaggerfallConnect.Arena2;
-using DaggerfallWorkshop.Utility;
+using Unity.Jobs;
+using Unity.Collections;
 
 namespace DaggerfallWorkshop
 {
@@ -33,12 +31,140 @@ namespace DaggerfallWorkshop
         const byte grass = 2;
         const byte stone = 3;
 
+        static readonly int tileDataDim = MapsFile.WorldMapTileDim + 1;
+
+        static readonly int assignTilesDim = MapsFile.WorldMapTileDim;
+
         byte[] lookupTable;
-        int[,] tileData;
 
         public TerrainTexturing()
         {
             CreateLookupTable();
+        }
+
+        public JobHandle ScheduleAssignTilesJob(ITerrainSampler terrainSampler, ref MapPixelData mapData, JobHandle dependencies, bool march = true)
+        {
+            // Cache tile data to minimise noise sampling during march.
+            NativeArray<byte> tileData = new NativeArray<byte>(tileDataDim * tileDataDim, Allocator.TempJob);
+            mapData.nativeArrayList.Add(tileData);
+            GenerateTileDataJob tileDataJob = new GenerateTileDataJob
+            {
+                heightmapData = mapData.heightmapData,
+                tileData = tileData,
+                tdDim = tileDataDim,
+                hDim = terrainSampler.HeightmapDimension,
+                maxTerrainHeight = terrainSampler.MaxTerrainHeight,
+                oceanElevation = terrainSampler.OceanElevation,
+                beachElevation = terrainSampler.BeachElevation,
+                mapPixelX = mapData.mapPixelX,
+                mapPixelY = mapData.mapPixelY,
+            };
+            JobHandle tileDataHandle = tileDataJob.Schedule(tileDataDim * tileDataDim, 64, dependencies);
+
+            // Assign tile data to terrain
+            NativeArray<byte> lookupData = new NativeArray<byte>(lookupTable, Allocator.TempJob);
+            mapData.nativeArrayList.Add(lookupData);
+            AssignTilesJob assignTilesJob = new AssignTilesJob
+            {
+                lookupTable = lookupData,
+                tileData = tileData,
+                tilemapData = mapData.tilemapData,
+                tdDim = tileDataDim,
+                tDim = assignTilesDim,
+                march = march,
+                locationRect = mapData.locationRect,
+            };
+            JobHandle assignTilesHandle = assignTilesJob.Schedule(assignTilesDim * assignTilesDim, 64, tileDataHandle);
+
+            return assignTilesHandle;
+        }
+
+        struct GenerateTileDataJob : IJobParallelFor
+        {
+            [ReadOnly]
+            public NativeArray<float> heightmapData;
+
+            public NativeArray<byte> tileData;
+
+            public int hDim;
+            public int tdDim;
+            public float maxTerrainHeight;
+            public float oceanElevation;
+            public float beachElevation;
+            public int mapPixelX;
+            public int mapPixelY;
+
+            // Gets noise value
+            private float NoiseWeight(float worldX, float worldY)
+            {
+                return GetNoise(worldX, worldY, 0.05f, 0.9f, 0.4f, 3, seed);
+            }
+
+            // Sets texture by range
+            private byte GetWeightedRecord(float weight, float lowerGrassSpread = 0.5f, float upperGrassSpread = 0.95f)
+            {
+                if (weight < lowerGrassSpread)
+                    return dirt;
+                else if (weight > upperGrassSpread)
+                    return stone;
+                else
+                    return grass;
+            }
+
+            // Noise function
+            private float GetNoise(
+                float x,
+                float y,
+                float frequency,
+                float amplitude,
+                float persistance,
+                int octaves,
+                int seed = 0)
+            {
+                float finalValue = 0f;
+                for (int i = 0; i < octaves; ++i)
+                {
+                    finalValue += Mathf.PerlinNoise(seed + (x * frequency), seed + (y * frequency)) * amplitude;
+                    frequency *= 2.0f;
+                    amplitude *= persistance;
+                }
+
+                return Mathf.Clamp(finalValue, -1, 1);
+            }
+
+            public void Execute(int index)
+            {
+                int x = JobA.Row(index, tdDim);
+                int y = JobA.Col(index, tdDim);
+
+                // Height sample for ocean and beach tiles
+                int hx = (int)Mathf.Clamp(hDim * ((float)x / (float)tdDim), 0, hDim - 1);
+                int hy = (int)Mathf.Clamp(hDim * ((float)y / (float)tdDim), 0, hDim - 1);
+                float height = heightmapData[JobA.Idx(hy, hx, hDim)] * maxTerrainHeight;  // x & y swapped in heightmap for TerrainData.SetHeights()
+                // Ocean texture
+                if (height <= oceanElevation)
+                {
+                    tileData[index] = water;
+                    return;
+                }
+                // Beach texture
+                // Adds a little +/- randomness to threshold so beach line isn't too regular
+                if (height <= beachElevation + (JobRand.Next(-15000, 15000) / 10000f))
+                {
+                    tileData[index] = dirt;
+                    return;
+                }
+
+                // Get latitude and longitude of this tile
+                int latitude = (int)(mapPixelX * MapsFile.WorldMapTileDim + x);
+                int longitude = (int)(MapsFile.MaxWorldTileCoordZ - mapPixelY * MapsFile.WorldMapTileDim + y);
+
+                // Set texture tile using weighted noise
+                float weight = 0;
+                weight += NoiseWeight(latitude, longitude);
+                // TODO: Add other weights to influence texture tile generation
+                tileData[index] = GetWeightedRecord(weight);
+            }
         }
 
         #region Marching Squares - WIP
@@ -46,90 +172,48 @@ namespace DaggerfallWorkshop
         // Very basic marching squares for water > dirt > grass > stone transitions.
         // Cannot handle water > grass or water > stone, etc.
         // Will improve this at later date to use a wider range of transitions.
-        public void AssignTiles(ITerrainSampler terrainSampler, ref MapPixelData mapData, bool march = true)
+        struct AssignTilesJob : IJobParallelFor
         {
-            // Cache tile data to minimise noise sampling
-            CacheTileData(terrainSampler, ref mapData);
+            [ReadOnly]
+            public NativeArray<byte> tileData;
+            [ReadOnly]
+            public NativeArray<byte> lookupTable;
 
-            // Assign tile data to terrain
-            int dim = MapsFile.WorldMapTileDim;
-            for (int y = 0; y < dim; y++)
+            public NativeArray<byte> tilemapData;
+
+            public int tdDim;
+            public int tDim;
+            public bool march;
+            public Rect locationRect;
+
+            public void Execute(int index)
             {
-                for (int x = 0; x < dim; x++)
+                int x = JobA.Row(index, tDim);
+                int y = JobA.Col(index, tDim);
+
+                // Do nothing if in location rect as texture already set, to 0xFF if zero
+                if (tilemapData[index] != 0)
+                    return;
+
+                // Assign tile texture
+                if (march)
                 {
-                    // Do nothing if location tile as texture already set
-                    if (mapData.tilemapSamples[x, y].location)
-                        continue;
+                    // Get sample points
+                    int tdIdx = JobA.Idx(x, y, tdDim);
+                    int b0 = tileData[tdIdx];               // tileData[x, y]
+                    int b1 = tileData[tdIdx + 1];           // tileData[x + 1, y]
+                    int b2 = tileData[tdIdx + tdDim];       // tileData[x, y + 1]
+                    int b3 = tileData[tdIdx + tdDim + 1];   // tileData[x + 1, y + 1]
 
-                    // Assign tile texture
-                    if (march)
-                    {
-                        // Get sample points
-                        int b0 = tileData[x, y];
-                        int b1 = tileData[x + 1, y];
-                        int b2 = tileData[x, y + 1];
-                        int b3 = tileData[x + 1, y + 1];
+                    int shape = (b0 & 1) | (b1 & 1) << 1 | (b2 & 1) << 2 | (b3 & 1) << 3;
+                    int ring = (b0 + b1 + b2 + b3) >> 2;
+                    int tileID = shape | ring << 4;
 
-                        int shape = (b0 & 1) | (b1 & 1) << 1 | (b2 & 1) << 2 | (b3 & 1) << 3;
-                        int ring = (b0 + b1 + b2 + b3) >> 2;
-                        int tileID = shape | ring << 4;
-
-                        byte val = lookupTable[tileID];
-                        mapData.tilemapSamples[x, y].record = val & 63;
-                        mapData.tilemapSamples[x, y].rotate = ((val & 64) == 64);
-                        mapData.tilemapSamples[x, y].flip = ((val & 128) == 128);
-                    }
-                    else
-                    {
-                        mapData.tilemapSamples[x, y].record = tileData[x, y];
-                    }
+                    tilemapData[index] = lookupTable[tileID];
                 }
-            }
-        }
-
-        void CacheTileData(ITerrainSampler terrainSampler, ref MapPixelData mapData)
-        {
-            // Create array if required
-            int dim = MapsFile.WorldMapTileDim + 1;
-            if (tileData == null)
-                tileData = new int[dim, dim];
-
-            // Populate array with tile metadata
-            for (int y = 0; y < dim; y++)
-            {
-                for (int x = 0; x < dim; x++)
+                else
                 {
-                    // Height sample for ocean and beach tiles
-                    float height = TerrainHelper.GetClampedHeight(
-                        ref mapData,
-                        terrainSampler.HeightmapDimension,
-                        (float)x / (float)dim,
-                        (float)y / (float)dim) * terrainSampler.MaxTerrainHeight;
-
-                    // Ocean texture
-                    if (height <= terrainSampler.OceanElevation)
-                    {
-                        tileData[x, y] = water;
-                        continue;
-                    }
-
-                    // Get latitude and longitude of this tile
-                    int latitude = (int)(mapData.mapPixelX * MapsFile.WorldMapTileDim + x);
-                    int longitude = (int)(MapsFile.MaxWorldTileCoordZ - mapData.mapPixelY * MapsFile.WorldMapTileDim + y);
-
-                    // Beach texture
-                    // Adds a little +/- randomness to threshold so beach line isn't too regular
-                    if (height <= terrainSampler.BeachElevation + UnityEngine.Random.Range(-1.5f, 1.5f))
-                    {
-                        tileData[x, y] = dirt;
-                        continue;
-                    }
-
-                    // Set texture tile using weighted noise
-                    float weight = 0;
-                    weight += NoiseWeight(latitude, longitude);
-                    // TODO: Add other weights to influence texture tile generation
-                    tileData[x, y] = GetWeightedRecord(weight);
+                    tilemapData[index] = tileData[JobA.Idx(x, y, tdDim)];
                 }
             }
         }
@@ -198,44 +282,6 @@ namespace DaggerfallWorkshop
             if (flip) index += 128;
 
             return (byte)index;
-        }
-
-        // Gets noise value
-        private float NoiseWeight(float worldX, float worldY)
-        {
-            return GetNoise(worldX, worldY, 0.05f, 0.9f, 0.4f, 3, seed);
-        }
-
-        // Sets texture by range
-        private int GetWeightedRecord(float weight, float lowerGrassSpread = 0.5f, float upperGrassSpread = 0.95f)
-        {
-            if (weight < lowerGrassSpread)
-                return dirt;
-            else if (weight > upperGrassSpread)
-                return stone;
-            else
-                return grass;
-        }
-
-        // Noise function
-        private float GetNoise(
-            float x,
-            float y,
-            float frequency,
-            float amplitude,
-            float persistance,
-            int octaves,
-            int seed = 0)
-        {
-            float finalValue = 0f;
-            for (int i = 0; i < octaves; ++i)
-            {
-                finalValue += Mathf.PerlinNoise(seed + (x * frequency), seed + (y * frequency)) * amplitude;
-                frequency *= 2.0f;
-                amplitude *= persistance;
-            }
-
-            return Mathf.Clamp(finalValue, -1, 1);
         }
 
         #endregion
