--- Assets/Scripts/Terrain/TerrainTexturingJobs.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f3a1a347943355d44974ee01c1a8e9b0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
